Move router and player-plugin logic out of the shared utils helper
