+++ src/components/PlayerGroupMembers.vue
@@ -46,7 +46,7 @@
 import PlayerIcon from "@/components/PlayerIcon.vue";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Separator } from "@/components/ui/separator";
-import { groupMemberPickerVisible } from "@/helpers/utils";
+import { groupMemberPickerVisible } from "@/helpers/players";
 import { api } from "@/plugins/api";
 import {
   type Player,
