+++ tests/components/PlayerCard.test.ts
@@ -80,6 +80,9 @@ vi.mock("@/helpers/utils", () => ({
       ? `${player.name} +${childCount}`
       : player.name;
   },
+}));
+
+vi.mock("@/helpers/players", () => ({
   isBuiltinPlayer: (player: Player) => player.player_id === "builtin",
 }));
